fix: split SkeletonNode into CurrentSkeletonNode and UpdatedSkeletonNode
