+++ crates/committer/src/patricia_merkle_tree.rs
@@ -1,6 +1,8 @@
+pub mod current_skeleton_node;
+pub mod current_skeleton_tree;
 pub mod errors;
 pub mod filled_node;
 pub mod filled_tree;
-pub mod skeleton_node;
-pub mod skeleton_tree;
 pub mod types;
+pub mod updated_skeleton_node;
+pub mod updated_skeleton_tree;
