+++ crates/committer/src/patricia_merkle_tree/types.rs
@@ -23,6 +23,12 @@ pub(crate) struct PathToBottom {
     pub length: EdgePathLength,
 }
 
+#[allow(dead_code)]
+pub(crate) struct EdgeData {
+    bottom_hash: HashOutput,
+    path_to_bottom: PathToBottom,
+}
+
 pub(crate) trait LeafDataTrait {
     /// Returns true if leaf is empty.
     fn is_empty(&self) -> bool;
